+++ docs/en/changes/changes.md
@@ -35,6 +35,7 @@
 * refactor: implement OTEL handler with SPI for extensibility.
 * chore: add `toString` implementation for `StorageID`.
 * chore: add a warning log when connecting to ES takes too long.
+* Fix the query time range in the metadata API.
 
 #### UI